This is a fast and dirty build system that adds PY VMODL files to VIB
It brute-force restarts vsanmgmtd on every start of our daemon.
It also assumes vsan service running - i.e. 'esxcli vsan cluster new' is the minimal requirement on 6.0 U2 (in U3 it does not require this command)

Also , the Makefiles are hacked. Basically, it is a change to unblock UI and testing, and it needs to be clean up
before merge to master